structs/unions: get rid of ManualDrop
Implement Clone, Copy and if possible, PartialEq and Debug on structs
and unions. Unfortunately, unions need specific implimentations of these
things and thus cannot derive it using the standard library. However,
the clone and copy can be derived since all the union elements are
defined behaviors (IIUC) in Rust. By doing this, it means that we can
avoid having any of these ManuallyDrop annotations on union types.

Signed-off-by: William Roberts <william.c.roberts@intel.com>